Move out code from Harness
