+++ tests/xharness/xharness.csproj
@@ -70,6 +70,7 @@
     <PackageReference Include="Mono.Options" Version="5.3.0.1" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AppInstallMonitorLog.cs" />
     <Compile Include="AppRunner.cs" />
     <Compile Include="BCLTestImporter\BCLTestAssemblyDefinition.cs" />
     <Compile Include="BCLTestImporter\BCLTestImportTargetFactory.cs" />
